Move _app and _error from routes/ssr to pages/ in root Remove unused fund routes
