+++ packages/social-trader/pages/_app.tsx
@@ -1,5 +1,33 @@
-import React from "react";
-import { _AppCreator } from "routes/ssr/_app";
-import { initializeStore } from "store";
+import "shared/styles/index.scss";
 
-export default _AppCreator(initializeStore);
+import withHistoryProvider from "decorators/history-provider/with-history-provider";
+import withReduxStore from "decorators/with-redux-store";
+import { appWithTranslation } from "i18n";
+import App from "next/app";
+import React from "react";
+import { Provider } from "react-redux";
+import { compose, Store } from "redux";
+import { InitializeStoreType } from "utils/types";
+
+export default (initializeStore: InitializeStoreType) => {
+  class CustomApp extends App<Props> {
+    render() {
+      const { Component, pageProps, reduxStore } = this.props;
+      return (
+        <Provider store={reduxStore}>
+          <Component {...pageProps} />
+        </Provider>
+      );
+    }
+  }
+
+  return compose(
+    withHistoryProvider,
+    appWithTranslation,
+    withReduxStore(initializeStore)
+  )(CustomApp);
+};
+
+interface Props {
+  reduxStore: Store;
+}
